chore: move secp256 crypto to primitive-traits
