+++ crates/primitives-traits/src/lib.rs
@@ -1,4 +1,17 @@
-//! Common abstracted types in Reth.
+//! Commonly used types and traits in Reth.
+//!
+//! This crate contains various primitive traits used across reth's components.
+//! It provides the [`Block`] trait which is used to represent a block and all its components.
+//! A [`Block`] is composed of a [`Header`] and a [`BlockBody`]. In ethereum (and optimism), a block body consists of a list of transactions, a list of uncle headers, and a list of withdrawals. For optimism, uncle headers and withdrawals are always empty lists.
+//!
+//! ## Feature Flags
+//!
+//! - `arbitrary`: Adds `proptest` and `arbitrary` support for primitive types.
+//! - `op`: Implements the traits for various [op-alloy](https://github.com/alloy-rs/op-alloy) types.
+//! - `reth-codec`: Enables db codec support for reth types including zstd compression for certain
+//!   types.
+//! - `serde`: Adds serde support for all types.
+//! - `secp256k1`: Adds secp256k1 support for transaction signing/recovery. (By default the no-std friendly `k256` is used)
 
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
@@ -40,6 +53,8 @@ pub use block::{
 mod encoded;
 mod withdrawal;
 pub use encoded::WithEncoded;
+
+pub mod crypto;
 
 mod error;
 pub use error::{GotExpected, GotExpectedBoxed};
