+++ crates/primitives-traits/src/transaction/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod execute;
 pub mod signed;
+pub mod signature;
 
 pub mod error;
