+++ crates/primitives-traits/src/transaction/signature.rs
@@ -1,0 +1,4 @@
+//! Signature types and helpers
+
+/// Re-exported signature type
+pub use alloy_primitives::PrimitiveSignature as Signature;
