+++ crates/primitives/src/transaction/signature.rs
@@ -2,15 +2,7 @@ use crate::transaction::util::secp256k1;
 use alloy_consensus::transaction::from_eip155_value;
 use alloy_primitives::{Address, PrimitiveSignature as Signature, B256, U256};
 use alloy_rlp::Decodable;
-
-/// The order of the secp256k1 curve, divided by two. Signatures that should be checked according
-/// to EIP-2 should have an S value less than or equal to this.
-///
-/// `57896044618658097711785492504343953926418782139537452191302581570759080747168`
-const SECP256K1N_HALF: U256 = U256::from_be_bytes([
-    0x7F, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-    0x5D, 0x57, 0x6E, 0x73, 0x57, 0xA4, 0x50, 0x1D, 0xDF, 0xE9, 0x2F, 0x46, 0x68, 0x1B, 0x20, 0xA0,
-]);
+use reth_primitives_traits::crypto::SECP256K1N_HALF;
 
 pub(crate) fn decode_with_eip155_chain_id(
     buf: &mut &[u8],
